# docs/api/api.yaml
openapi: "3.0.0"
info:
  title: "Wutsi SMS API"
  description: |
    Wutsi SMS Server. It provides the following features:
    - Sending SMS
    - Verifying SMS number
  version: "1.0.0"

servers:
  - url: https://wutsi-sms-server-test.herokuapp.com
    description: Sandbox
  - url: https://wutsi-sms-server-prod.herokuapp.com
    description: Production

paths:
  /v1/sms/messages:
    post:
      operationId: "send-message"
      description: Sends a SMS Message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendMessageResponse'
        409:
          description: SMS delivery failed
      tags:
        - Message
      security:
        - api_key: [ 'sms-send' ]

  /v1/sms/verifications:
    post:
      operationId: "send-verification"
      description: Send a verification message to a SMS number. Each verification sent expires in 5 minutes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendVerificationRequest'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendVerificationResponse'
        409:
          description: SMS delivery failed
      tags:
        - Verification
      security:
        - api_key: [ 'sms-verify' ]

  /v1/sms/verifications/{id}:
    get:
      operationId: "validate-verification"
      description: Validate a verification
      parameters:
        - in: path
          name: id
          description: ID of the verification request
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: code
          description: verification code
          required: true
          schema:
            type: string
      responses:
        200:
          description: Validation is successfull
        409:
          description: |
            Validation failed failed. The verification will fail when
            - The ID of the verification request is invalid
            - The code provided is invalid
            - The verification has expired
            - The verification has already been verified
      tags:
        - Verification
      security:
        - api_key: [ 'sms-verify' ]


components:
  securitySchemes:
    api_key:
      type: apiKey
      in: header
      name: Authorization

  schemas:
    SendMessageRequest:
      type: object
      properties:
        phoneNumber:
          type: string
          description: |
            Phone number in the [E164](https://en.wikipedia.org/wiki/E.164) format
        message:
          type: string
          description: Message to send. A SMS message can contains up to 160 characters
          maxLength: 160
      required:
        - phoneNumber
        - message

    SendMessageResponse:
      type: object
      properties:
        id:
          type: string
          description: Message unique identifier

    SendVerificationRequest:
      type: object
      properties:
        phoneNumber:
          type: string
          description: |
            Phone number in the [E164](https://en.wikipedia.org/wiki/E.164) format
        language:
          type: string
          description: |
            2 letter code of the language ([ISO 659-1](https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes)).
            Ex: `en`, `fr`
          default: en
          maxLength: 2
      required:
        - number

    SendVerificationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Verification request identifier